Initial commit
This library is designed to make it easier to test JSON response data
against schema definitions, to match the structure and value types of
the data and provide assurance that the API responses do not change
unexpectedly.

Schema definitions can be composed and recomposed very fluidly, making
it easy to avoid having to redefine the schemas for commonly-nested
objects over and over again.

It also, probably more importantly, provides a mechanism to guarantee
that the underlying schema structure being used in a test cannot change
without some evidence of that change in the test iself. An unfortunate
consequence of avoiding repetition within schema definitions and between
test implementations is that the _impact_ of schema changes may not be
evident, but this library makes it _impossible_ to impact a test without
there being some visible diff in the relevant assertions.