Add spec for implicit versioning
If no version is given, but the containing definition has a version,
then try to look up the object using that container definition version.

This does mean that if no versioned object definition exists, the
containing definition will failed. Perhaps we should fall back to object
definitions with `nil` version?