Handle rewriting over previously expanded versions
We can't assume the rewriter is always replacing a single-line
definition with an expanded one; doing so makes the use of
`current_offset` immediately wrong as soon as we encounter a
previously-expanded definition.

Instead, try to track the current definition length and calculate the
offset as the difference between that length and the new one.