Allow attributes to be nullable
Sometimes there's no easy way to avoid testing against a null value --
such as in pagination.

While doing this, I noticed that it was not possible to call `optional`
in sub-definitions of objects, because our `attr_reader` was being used
as the implementation for `optional` (and `nullable`) in these cases.
The solution is to move the definitional `optional` and `nullable`
methods up into the top-level `Property` class, from which `Definition`
inherits, so that all objects can have access to this mechanism.