Allow setting of object and array types
In order to be able to define the custom schema types for an array of
objects, we need to be able to call `type` as a method, even within an
object definition. This means we need to raise that method up into the
Property superclass, and move the attr_reader `type` method out of the
way (TBD, a better name for it).