Allow duplicate callsites if result matches schema
Previously if assert_valid_json_for_schema was called in a loop (e.g.
tests/contexts defined dynamically), schema-test would try to rewrite
the same callsite multiple times, which works the first time but fails
afterwards because the file was already updated.

Now we check if we visited this site before and quietly continue if the
expected result is the same. If the result is different between the
assertions we raise to inform the user that their expectation performs
differently within each loop.