Handle offset in rubocop-disabled code
The rewriter works by trying to find and delete any rubocop lines
adjacent to the schema assertion, then writes them back at the end if
the `disable_rubocop` option is true.

Assuming we delete the rubocop magic comments, the file we operate on
becomes 2 lines shorter, so we need to subtract this from the offset.

I've added some tests to assert this case, which requires at least 2
schema assertions in the same file to be validated.